PATCH: Fix NMP build of `@neos-project/neos-ui-extensibility` by separating types declaration
otherwise typescript attempts to parse the whole neos ui repo and that fails
